Extract bypass-lineage tracking into a mixin (WS5 seam)
Move the 5 methods that record and hand off which sessions descend
from an LCM-bypassed session (_has/_mark/_unmark_lcm_bypass_lineage_
session, _handoff_lcm_bypass_lineage, _compression_boundary_from_lcm_
bypassed_session) verbatim out of engine.py into a new
BypassLineageMixin in bypass_lineage.py. LCMEngine mixes it in before
ContextEngine, so `self` still resolves every reference (including the
shared auxiliary-session lock); call sites, shared state, and method
names are unchanged.

Behaviour-neutral: the moved method bodies are byte-identical to main
(AST-slice compare, 0 diffs). ruff clean; full suite 1575 passed /
12 xfailed (unchanged from the v0.19.0 baseline).
